fix(bitfields): Improve bitfield handling
Compilers pack bitfields to smaller than the type if they fit in.
For example, if only one bit is used in a bitfield of uint32_t
then the pack struct will fit it in a uint8_t making the memory
offset incorrect.
This fills in reserved bitfields so the structure is full
preventing the compiler from making unwanted optimizations.

Also removes the bitfield type from the CSV which helps with
autocomplete not require a . for the bits.
Version bump since this will be released.
Also breaks out some generated c data with newlines to make
it a bit better for git.
Regression tests updated.